revert: remove mdx-loader.ts, restore inline import.meta.glob
Extracting import.meta.glob to a shared module caused hydration
mismatch — SSR and client Vite environments compiled MDX differently.
Restore inline globs in entry-client.tsx (client-only) and source.ts
(SSR-only) to keep each environment's compilation isolated.
